+++ src/Tests/ContainerGroup.fs
@@ -2,6 +2,7 @@ module ContainerGroup
 
 open Expecto
 open Farmer
+open Farmer.Arm
 open Farmer.Identity
 open Farmer.ContainerGroup
 open Farmer.Arm.ContainerInstance
@@ -626,5 +627,92 @@ async {
         let zones = jobj.SelectToken "resources[?(@.name=='zonal-container-group')].zones"
         Expect.hasLength zones 1 "Incorrect number of zones"
         Expect.sequenceEqual zones [JValue "2"] "Incorrect value for zone"
+    }
+
+    test "Enable container logging workspace" {
+        let deployment =
+            let workspace = logAnalytics { name "containergrouplogs1234" }
+            arm {
+                add_resources [
+                    workspace
+                    containerGroup {
+                        name "container-group-with-insights"
+                        add_instances [
+                            containerInstance {
+                                name "httpserver"
+                                image "nginx"
+                            }
+                        ]
+                        diagnostics_workspace LogType.ContainerInstanceLogs workspace
+                    }
+                ]
+            }
+        let jobj = deployment.Template |> Writer.toJson |> JObject.Parse
+        let logAnalytics = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].properties.diagnostics.logAnalytics"
+        let workspaceId = logAnalytics.SelectToken "workspaceId"
+        let workspaceKey = logAnalytics.SelectToken "workspaceKey"
+        let logType = logAnalytics.SelectToken "logType"
+        Expect.equal (string workspaceId) "[reference(resourceId('Microsoft.OperationalInsights/workspaces', 'containergrouplogs1234'), '2020-03-01-preview').customerId]" "Incorrect value for workspaceId"
+        Expect.equal (string workspaceKey) "[listkeys(resourceId('Microsoft.OperationalInsights/workspaces', 'containergrouplogs1234'), '2020-03-01-preview').primarySharedKey]" "Incorrect value for workspaceKey"
+        Expect.equal (string logType) "ContainerInstanceLogs" "Incorrect value for workspaceId"
+        let cgDependencies = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].dependsOn"
+        Expect.hasLength cgDependencies 1 "Incorrect number of dependencies for diagnostics workspace"
+    }
+
+    test "Enable linking to container logging workspace" {
+        let deployment =
+            let workspaceId = LogAnalytics.workspaces.resourceId "my-log-analytics-workspace"
+            arm {
+                add_resources [
+                    containerGroup {
+                        name "container-group-with-insights"
+                        add_instances [
+                            containerInstance {
+                                name "httpserver"
+                                image "nginx"
+                            }
+                        ]
+                        link_to_diagnostics_workspace LogType.ContainerInstanceLogs workspaceId
+                    }
+                ]
+            }
+        let jobj = deployment.Template |> Writer.toJson |> JObject.Parse
+        let logAnalytics = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].properties.diagnostics.logAnalytics"
+        let workspaceId = logAnalytics.SelectToken "workspaceId"
+        let workspaceKey = logAnalytics.SelectToken "workspaceKey"
+        let logType = logAnalytics.SelectToken "logType"
+        Expect.equal (string workspaceId) "[reference(resourceId('Microsoft.OperationalInsights/workspaces', 'my-log-analytics-workspace'), '2020-03-01-preview').customerId]" "Incorrect value for workspaceId"
+        Expect.equal (string workspaceKey) "[listkeys(resourceId('Microsoft.OperationalInsights/workspaces', 'my-log-analytics-workspace'), '2020-03-01-preview').primarySharedKey]" "Incorrect value for workspaceKey"
+        Expect.equal (string logType) "ContainerInstanceLogs" "Incorrect value for workspaceId"
+        let cgDependencies = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].dependsOn"
+        Expect.isEmpty cgDependencies "Should have no dependencies when linking to a workspace."
+    }
+
+    test "Enable passing key to container logging workspace" {
+        let fakeWorkspaceId = Guid.NewGuid() |> string
+        let fakeWorkspaceKey = Guid.NewGuid() |> string
+        let deployment =
+            arm {
+                add_resources [
+                    containerGroup {
+                        name "container-group-with-insights"
+                        add_instances [
+                            containerInstance {
+                                name "httpserver"
+                                image "nginx"
+                            }
+                        ]
+                        diagnostics_workspace_key LogType.ContainerInstanceLogs fakeWorkspaceId fakeWorkspaceKey
+                    }
+                ]
+            }
+        let jobj = deployment.Template |> Writer.toJson |> JObject.Parse
+        let logAnalytics = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].properties.diagnostics.logAnalytics"
+        let workspaceId = logAnalytics.SelectToken "workspaceId"
+        let workspaceKey = logAnalytics.SelectToken "workspaceKey"
+        Expect.equal (string workspaceId) fakeWorkspaceId "Incorrect value for workspaceId"
+        Expect.equal (string workspaceKey) fakeWorkspaceKey "Incorrect value for workspaceKey"
+        let cgDependencies = jobj.SelectToken "resources[?(@.name=='container-group-with-insights')].dependsOn"
+        Expect.isEmpty cgDependencies "Should have no dependencies when linking to a workspace."
     }
 ]